Safari-609.1.11.1/LayoutTests/http/tests/security/window-name-after-cross-origin-aux-frame-navigation-expected.txt

CONSOLE MESSAGE: line 10: window.name should equal "test" after a cross-origin auxiliary frame navigation.
CONSOLE MESSAGE: line 4: window.name equals "test".
